+++ inspectors/apt/packages_test.go
@@ -143,22 +143,25 @@ var packagesInspectArtifactTests = []packagesInspectArtifactTest{{
 func (s *aptSuite) TestPackagesInspectArtifact(c *C) {
 	for _, tc := range packagesInspectArtifactTests {
 		ins := apt.NewAptPackagesInspector(getAptInspectorConfig())
+		h2, _ := digests.NewSha256Digest(tc.digest)
 
 		// simulate InRelease entry
-		data := apt.NewAptPackages("http://myserver", "jammy", "main", "amd64")
+		data := apt.NewAptPackages(h2, "http://myserver", "jammy", "main", "amd64")
 		apt.AptPackagesInspectorAddPackages(ins, "http://myserver", "/path/Packages.xz", data, s.slog)
-
-		h2, _ := digests.NewSha256Digest(tc.digest)
 
 		a := metadata.NewArtifact()
 		a.CurrentDownload.URL = "http://myserver/path/Packages.xz"
 		a.Metadata.Sha256 = h2
 		a.Metadata.Type = "application/x.apt.packages"
 		a.MimeType = mimetype.Lookup("application/x.apt.packages")
-		a.RequestInspection[ins.ID()] = &Inspection{
-			Opinion: opinions.Pending,
-			Reason:  "some reason",
-		}
+		a.SetRequestPending(ins, "some reason").Annotate(
+			Annotation{
+				"cfg-name":     "default",
+				"dist":         "jammy",
+				"component":    "main",
+				"architecture": "amd64",
+			},
+		)
 		a.ResponseInspection["apt.release"] = &Inspection{
 			Opinion: opinions.Unknown,
 			Reason:  "some reason",
@@ -246,6 +249,9 @@ func (s *aptSuite) TestPackageParsing(c *C) {
 }
 
 type aptPackagesDebValidationTest struct {
+	pkgUrl            string // Packages file URL
+	debUrl            string // Deb file URL
+	cfgName           string // The repository name in configuration
 	filename          string // Test artifact file name
 	debInPackages     bool   // Whether the deb is listed in the Packages file
 	packagesInRelease bool   // Whether the Packages file is listed in InRelease
@@ -253,20 +259,45 @@ type aptPackagesDebValidationTest struct {
 }
 
 var aptPackagesDebValidationTests = []aptPackagesDebValidationTest{{
+	pkgUrl:            "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513",
+	debUrl:            "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
+	cfgName:           "default",
 	filename:          "testdata/2048_0.20210105.1243-1_amd64.deb",
 	debInPackages:     true,
 	packagesInRelease: true,
 	rejectReason:      "",
 }, {
+	pkgUrl:            "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513",
+	debUrl:            "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
+	cfgName:           "default",
 	filename:          "testdata/2048_0.20210105.1243-1_amd64.deb",
 	debInPackages:     true,
 	packagesInRelease: false,
 	rejectReason:      "artifact listed in invalid Packages file",
 }, {
+	pkgUrl:            "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513",
+	debUrl:            "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
+	cfgName:           "default",
 	filename:          "testdata/2048_0.20210105.1243-1_amd64.deb",
 	debInPackages:     false,
 	packagesInRelease: true,
 	rejectReason:      "deb file digest not listed in packages file",
+}, {
+	pkgUrl:            "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513",
+	debUrl:            "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
+	cfgName:           "", // missing cfg-name
+	filename:          "testdata/2048_0.20210105.1243-1_amd64.deb",
+	debInPackages:     true,
+	packagesInRelease: true,
+	rejectReason:      "deb file downloaded from unknown repository",
+}, {
+	pkgUrl:            "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513",
+	debUrl:            "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
+	cfgName:           "invalid", // invalid cfg-name
+	filename:          "testdata/2048_0.20210105.1243-1_amd64.deb",
+	debInPackages:     true,
+	packagesInRelease: true,
+	rejectReason:      "Unknown repository configuration name",
 }}
 
 func (s *aptSuite) TestAptPackagesDebValidation(c *C) {
@@ -279,7 +310,7 @@ func (s *aptSuite) TestAptPackagesDebValidation(c *C) {
 		// Inspect the packages file
 		if tc.debInPackages {
 			a2 := metadata.NewArtifact()
-			a2.CurrentDownload.URL = "http://archive.ubuntu.com/ubuntu/dists/jammy/main/binary-amd64/by-hash/SHA256/5bf8bcda46d2d2bd81840ea5808f2b2122598bd68cc4889e3e32055ecc43b513"
+			a2.CurrentDownload.URL = tc.pkgUrl
 			err := ins.InspectRequest(a2)
 			c.Assert(err, IsNil)
 			c.Assert(a2.RequestInspection[ins.ID()].Opinion, Equals, opinions.Pending)
@@ -307,11 +338,14 @@ func (s *aptSuite) TestAptPackagesDebValidation(c *C) {
 		a := metadata.NewArtifact()
 		a.Metadata.Type = "application/vnd.debian.binary-package"
 		a.MimeType = mimetype.Lookup("application/vnd.debian.binary-package")
-		a.CurrentDownload = metadata.Download{
-			URL: "http://archive.ubuntu.com/ubuntu/pool/universe/2/2048/2048_0.20210105.1243-1_amd64.deb",
-		}
+		a.CurrentDownload = metadata.Download{URL: tc.debUrl}
 		a.Metadata.Sha256, _ = digests.NewSha256Digest("defa3a9b60849dc9f1bd3549381683d9fd245a0bebb900dc84f306c133a05a17")
 		a.Metadata.Size = 14744
+		notes := Annotation{}
+		if tc.cfgName != "" {
+			notes["cfg-name"] = tc.cfgName
+		}
+		a.SetRequestPending(ins, "some reason").Annotate(notes)
 
 		pkgreader, err := os.Open(tc.filename)
 		c.Assert(err, IsNil)