+++ inspectors/lxd/urls.go
@@ -27,6 +27,8 @@ import (
 
 // Recognized URL format:
 // ----------------------
+// https://cloud-images.ubuntu.com:443/buildd/daily/streams/v1/index.json
+// https://cloud-images.ubuntu.com:443/buildd/daily/streams/v1/com.ubuntu.cloud:daily:download.json
 // http://cloud-images.ubuntu.com/buildd/daily/noble/20250629/noble-server-cloudimg-amd64-lxd_combined.tar.gz
 // https://cloud-images.ubuntu.com:443/buildd/daily/noble/20250629/noble-server-cloudimg-amd64-lxd_combined.tar.gz
 
@@ -36,9 +38,12 @@ var (
 		regexp.MustCompile(`^https://cloud-images.ubuntu.com:443$`),
 	}
 
-	reSimpleStreamsIndex = regexp.MustCompile(`^/([\w-\/]+)/streams/v1/index.json$`)
-	reRootfs             = regexp.MustCompile(`^/buildd/daily/([\w-]+)/(\d+)/([\w+\.-]+\.tar.gz)$`)
+	reSimpleStreamsIndex    = regexp.MustCompile(`^/([\w-\/]+)/streams/v1/index.json$`)
+	reSimpleStreamsDownload = regexp.MustCompile(`^/([\w-\/]+)/streams/v1/([\w-\.\/:]+):download.json$`)
+	reProductItem           = regexp.MustCompile(`^/buildd/(daily|releases)/([\w-]+)/([\w-]+)/([\w+\.-]+\.tar.gz)$`)
 )
+
+// Simple streams index
 
 type SimpleStreamsIndexUrlInfo struct {
 	Stream string // The stream portion of the URL
@@ -60,26 +65,54 @@ func NewSimpleStreamsIndexUrlInfo(u *url.URL) (*SimpleStreamsIndexUrlInfo, error
 	return info, nil
 }
 
-type RootfsUrlInfo struct {
-	Series string // The image series
-	Date   string // The date of the daily image
-	Name   string // The rootfs filename
+// Simple stream product download request
+
+type SimpleStreamsDownloadUrlInfo struct {
+	Stream   string // The stream portion of the URL
+	ItemPath string // The image to download
 }
 
-func NewRootfsUrlInfo(u *url.URL) (*RootfsUrlInfo, error) {
+func NewSimpleStreamsDownloadUrlInfo(u *url.URL) (*SimpleStreamsDownloadUrlInfo, error) {
 	if err := checkValidOrigin(u); err != nil {
 		return nil, err
 	}
 
-	m := reRootfs.FindStringSubmatch(u.Path)
-	if len(m) != 4 {
+	m := reSimpleStreamsDownload.FindStringSubmatch(u.Path)
+	if len(m) != 3 {
 		return nil, fmt.Errorf("invalid URL path: %s", u.Path)
 	}
 
-	info := &RootfsUrlInfo{
-		Series: m[1],
-		Date:   m[2],
-		Name:   m[3],
+	info := &SimpleStreamsDownloadUrlInfo{
+		Stream:   m[1],
+		ItemPath: m[2],
+	}
+	return info, nil
+}
+
+// LXD product tarball
+
+type ProductItemUrlInfo struct {
+	ImageType string // Daily or releases
+	Series    string // The image series
+	Date      string // The date of the daily image
+	Name      string // The rootfs filename
+}
+
+func NewProductItemUrlInfo(u *url.URL) (*ProductItemUrlInfo, error) {
+	if err := checkValidOrigin(u); err != nil {
+		return nil, err
+	}
+
+	m := reProductItem.FindStringSubmatch(u.Path)
+	if len(m) != 5 {
+		return nil, fmt.Errorf("invalid URL path: %s", u.Path)
+	}
+
+	info := &ProductItemUrlInfo{
+		ImageType: m[1],
+		Series:    m[2],
+		Date:      m[3],
+		Name:      m[4],
 	}
 	return info, nil
 }